tests/requirements.txt
djangocms-helper @ git+https://github.com/nephila/django-app-helper.git@3.3.1
tox>=2.9.1
coverage>=4.4.2
django-polymorphic
djangocms-text-ckeditor>=3.7.0
captcha>=0.4
django-import-export>=2.2.0
django-treebeard
flake8
isort
babel
